# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OntoloSQL Designer Desktop API",
    "version": "1.4.0",
    "description": "Versioned API for native desktop clients. Web SPA authentication remains Sanctum cookie/session based; this API uses Google OAuth with PKCE and expiring Sanctum bearer tokens.\n\nVersion 1.1.0 documents ontology value types in detail, including enum value types (a string base type with a `oneOf` constraint), and the visual-only schema markers used for reference tables, reference links, pipeline transforms, and table-level Foundry edit settings (`editsEnabled` / `editsHistory`)."
  },
  "servers": [
    {
      "url": "https://ontolosql-designer-kripb.ondigitalocean.app",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local Docker development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/diagrams": {
      "get": {
        "tags": [
          "Diagrams"
        ],
        "operationId": "listOwnedDiagramsV1",
        "summary": "List current user's diagrams",
        "description": "Requires abilities: desktop, diagrams:read.",
        "responses": {
          "200": {
            "description": "Pageless diagram summary collection",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DiagramSummary"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Diagrams"
        ],
        "operationId": "createDiagramV1",
        "summary": "Create diagram",
        "description": "Requires abilities: desktop, diagrams:write.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiagramRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/DiagramCreated"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
    "/api/v1/foundry/connections": {
      "get": {
        "tags": [
          "Foundry"
        ],
        "operationId": "listFoundryConnections",
        "summary": "List my Foundry connections",
        "description": "Requires abilities: desktop, foundry:connect.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/FoundryConnection"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/foundry/connections/{connection}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/FoundryConnectionId"
        }
      ],
      "delete": {
        "tags": [
          "Foundry"
        ],
        "operationId": "deleteFoundryConnection",
        "summary": "Disconnect a Foundry connection",
        "description": "Requires abilities: desktop, foundry:connect. Only the owning user may disconnect.",
        "responses": {
          "204": {
            "description": "Disconnected"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum"
      }
    },
    "parameters": {
...
      "FoundryConnectionId": {
        "name": "connection",
        "in": "path",
        "required": true,
        "description": "Foundry connection id owned by the authenticated user.",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
...
    },
...
    "responses": {
      "DiagramCreated": {
        "description": "Diagram created",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "diagram"
              ],
              "properties": {
                "status": {
                  "type": "boolean"
                },
                "message": {
                  "type": "string"
                },
                "diagram": {
                  "type": "object",
                  "required": [
                    "id",
                    "share_token"
                  ],
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "share_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "Authenticated token lacks the required ability or diagram permission",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Missing, invalid, or expired bearer token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "CreateDiagramRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255
          },
          "db_type": {
            "$ref": "#/components/schemas/DbType"
          },
          "share_access": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiagramAccess"
              },
              {
                "type": "null"
              }
            ]
          },
          "library": {
            "type": "boolean"
          }
        }
      },
...
      "DbType": {
        "type": "string",
        "enum": [
          "mysql",
          "postgresql",
          "sqlite",
          "oracle",
          "sqlserver",
          "msaccess",
          "ontology"
        ]
      },
...
      "DiagramAccess": {
        "type": "string",
        "enum": [
          "read",
          "write",
          "per_user"
        ]
      },
      "DiagramSchema": {
        "type": "array",
        "description": "Vue Flow schema elements (a flat list of nodes and edges). Nodes have a `type` of `table`, `row`, or `pipeline-transform`; edges have `source`/`target` referencing node ids.\n\nNotable `data` fields: table nodes may carry `ontologyActions`, `editsEnabled` (boolean; Foundry object patches enabled) and `editsHistory` (`{ enabled, storeAllPreviousProperties }`), `titlePropertyRowId`, and `implementsInterfaces`. Row nodes carry `sqlType`, `keyMod`, `nullable`, `indexed`, `userEdits`, and `valueTypeId` (references a ValueType `id`).\n\nVisual-only / reference elements are marked and excluded from SQL and Maker exports but preserved in backup JSON: reference tables (`data.tableKind = 'reference'`, `data.reference = true`), reference links (`data.linkKind = 'reference'`), pipeline transform nodes (`type = 'pipeline-transform'`), and transform edges (`type = 'transform'`, `data.linkKind = 'transform'`). Any element with `data.exportable = false` is visual-only.\n\nRuntime-only keys such as hidden, selected, dimensions, and editing are not persisted and are stripped from save/load paths.",
        "items": {
          "type": "object",
          "additionalProperties": true
        }
      },
...
      "DiagramSummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "db_type",
          "schema",
          "share_token",
          "require_approval",
          "library",
          "is_owner"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "db_type": {
            "$ref": "#/components/schemas/DbType"
          },
          "schema": {
            "$ref": "#/components/schemas/DiagramSchema"
          },
          "share_token": {
            "type": "string"
          },
          "share_access": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiagramAccess"
              },
              {
                "type": "null"
              }
            ]
          },
          "effective_access": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiagramAccess"
              },
              {
                "type": "null"
              }
            ]
          },
          "require_approval": {
            "type": "boolean"
          },
          "library": {
            "type": "boolean"
          },
          "is_owner": {
            "type": "boolean"
          },
          "owner_email": {
            "type": [
              "string",
              "null"
            ],
            "format": "email"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "FoundryConnection": {
        "type": "object",
        "description": "A user's delegated OAuth connection to a Foundry host. Tokens are never exposed.",
        "properties": {
          "id": {
            "type": "integer"
          },
          "host_url": {
            "type": "string",
            "format": "uri"
          },
          "display_name": {
            "type": "string",
            "nullable": true
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "active": {
            "type": "boolean"
          },
          "revoked": {
            "type": "boolean"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "last_used_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "auth_type": {
            "type": "string",
            "enum": [
              "oauth",
              "token"
            ],
            "description": "How the connection was established."
          }
        },
        "required": [
          "id",
          "host_url",
          "active"
        ]
      },
...
    }
  }
}